[iOS] Finchable width for NTP containment
Updates HomeModuleMinimumPadding param to use a percentage width instead of a fixed margin (e.g. 92% instead of 24pt).

Internal CL to update feed content width: cl/623234757

Bug: 333398611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5440233
Commit-Queue: Adam Arcaro <adamta@google.com>
Auto-Submit: Adam Arcaro <adamta@google.com>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1285173}

+++ ios/chrome/browser/shared/public/features/features.h
@@ -545,8 +545,9 @@ extern const char kTabResumptionAllTabsOneDayThresholdParam[];
 bool IsFeedContainmentEnabled();
 
 // The minimum padding between the modules and the screen bounds on the Home
-// surface. Relies on `IsFeedContainmentEnabled()` being enabled.
-int HomeModuleMinimumPadding();
+// surface. Relies on `IsFeedContainmentEnabled()` being enabled. This padding
+// is dynamic, so the value represents a percentage including both sides.
+CGFloat HomeModuleMinimumPadding();
 
 // Whether the tab resumption feature is enabled.
 bool IsTabResumptionEnabled();
